Add files via upload
Write a JavaScript which will be used in an HTML document that contains a table having at least 3 columns and several lines. When the user clicks the table header the script will sort in ascending order the elements from the current column. If the user clicks again on the column header, the elements will be sorted in descending order.